bugfixes, methods updates, tests

+++ include/libtomo/stripe.h
@@ -58,13 +58,15 @@ DLL int
 stripesdetect3d_main_float(float* Input, float* Output, 
                            int window_halflength_vertical,
                            int ratio_radius,
-                           int ncores, int dimX, int dimY, int dimZ);
+                           int ncores,
+                           long dimX, long dimY, long dimZ);
 
 DLL int
-stripesmask3d_main_float(float* Input, unsigned char* Output,
-                         float threshold_val,
-                         int stripe_length_min,
-                         int stripe_depth_min,
-                         int stripe_width_min,
-                         float sensitivity,
-                         int ncores, int dimX, int dimY, int dimZ);
+stripesmask3d_main_float(float* Input,
+                        unsigned char* Output,
+                        float threshold_val,
+                        int stripe_length_min,
+                        int stripe_depth_min,
+                        int stripe_width_min,
+                        float sensitivity,
+                        int ncores, long dimX, long dimY, long dimZ);
